Add notion of isDismissing and gate onReady() callbacks on it (#243)
* Add notion of isDismissing and gate onReady() callbacks on it

There was a race condition I came across where a view could be animating out at the same time as the animation for dismissing the tap target. If the view animated out before the taptarget, it would trigger a layout again, and potentially redisplay the taptarget again. This gates onReady callbacks against the notion of "isDismissing". Tested locally in my project and confirmed this resolves the issue.

* Return early (and earlier!)